+++ TrelloDotNet/TrelloDotNet.Tests/FactManualOnlyAttribute.cs
@@ -1,0 +1,9 @@
+﻿namespace TrelloDotNet.Tests;
+
+public sealed class FactManualOnlyAttribute : FactAttribute
+{
+    public FactManualOnlyAttribute()
+    {
+        Skip = "Manual";
+    }
+}